Snapshot of kokkos.git from commit c1c9e8d6b30e58b04700bb4602494bb417d42fbc
From repository at git@github.com:/kokkos/kokkos.git

At commit:
commit c1c9e8d6b30e58b04700bb4602494bb417d42fbc
Author: crtrott <crtrott@sandia.gov>
Date:   Tue Dec 8 23:06:55 2015 -0700

    Core: fix another issue with the memory tracking, introduced recently

    There was a bug introduced when solving the performance issues with
    subviews. It seemed to not have trickered any errors in either Trilinos
    tests nor kokkos tests, but did crash Nalu.

+++ lib/kokkos/core/src/impl/Kokkos_Serial_TaskPolicy.cpp
@@ -230,7 +230,7 @@ void Task::assign( Task ** const lhs , Task * rhs , const bool no_throw )
 namespace {
 
 Task * s_ready = 0 ;
-Task * s_denied = reinterpret_cast<Task*>( ~((unsigned long)0) );
+Task * s_denied = reinterpret_cast<Task*>( ~((uintptr_t)0) );
 
 }
